diag(gc): make PERRY_STACKMAP_WALKER=verify name the disagreement it finds
The first end-to-end `verify` run on aarch64 ELF caught the fp-chain walker
and the Itanium unwinder resolving one root 96 bytes apart (#7984). All the
gate could say was "1 unique slot versus 1 unique slot" and print the two
addresses in decimal — from which the frame, the base register, the function
whose prologue was decoded, and therefore *which walker is wrong* are all
unrecoverable. Every candidate explanation (a missed trailing `sub sp`, a
frame the chain skipped, a CFA one frame out) predicts exactly that output.

Both walkers now report a `ResolvedRoot` instead of a bare `MutableRootSlot`:
the same address, plus the frame return address it was matched on, the record's
function, the map's base register and frame offset, and the base that walker
resolved the register to. `verify` prints all of it on a mismatch, calls out an
equal-slot-count disagreement as a base disagreement rather than a missed frame,
and on aarch64 dumps `fp_to_sp_offset`'s decode and the prologue words it read.

The prologue dump is gated on the map vouching for the function address. The
first draft was not, and a unit test with a synthetic address turned the
diagnostic into a SIGSEGV with no output — which is what would also happen in
the field for the failure mode "the map's addresses are wrong".

`verify` moves to its own file and the decoder tests to another, because
`stack_maps.rs` was eight lines under the 2000-line cap.

The workflow's crash path tailed 20 lines of the failing run's stderr, which
truncated the report's head. 120.
